\name{lxplot}
\alias{lxplot}
\title{Log-X Plot of Spectrum}
\description{
  Log-X Plot of Spectrum}
}
\usage{
lxplot(yper, spec)
}
\arguments{
  \item{yper}{periodogram values}
  \item{spec}{spectrum values}
}
\author{Jan Beran (principal) and Martin Maechler (speedup, fine tuning)}
%\seealso{ ~~objects to SEE ALSO as \code{\link{~~fun~~}}, ~~~ }
%\examples{
%}
\keyword{hplot}
\keyword{ts}
