\name{llplot}
\alias{llplot}
\title{Log-Log Plot of Spectrum}
\description{
  Log-Log Plot of Spectrum}
}
\usage{
llplot(yper, spec)
}
\arguments{
  \item{yper}{periodogram values}
  \item{spec}{spectrum values}
}
%\value{
% }
\author{Jan Beran (principal) and Martin Maechler (speedup, fine tuning)}
%\seealso{ ~~objects to SEE ALSO as \code{\link{~~fun~~}}, ~~~ }
%\examples{
%}
\keyword{hplot}
\keyword{ts}
