\name{NBSdiff1kg}
\alias{NBSdiff1kg}
\docType{data}
\title{NBS measurement deviations from 1 kg}
\description{
  NBS weight measurements - deviation from 1 kg in micrograms
  (Pollak, Croarkin and Hagwood 1993, Graf, Hampel and Tacier 1984,
  Graf 1983). The data are listed in chronological sequence by row.
}
\usage{data(NBSdiff1kg)}
\format{
 A time series of length 289.
}
\source{Jan Beran and Brandon Whitcher by E-mail in fall 1995.
}
\references{
  %% FIXME ---
}
\examples{
data(NBSdiff1kg)
## maybe str(NBSdiff1kg) ; plot(NBSdiff1kg) ...
}
\keyword{datasets}
